chore(tests): split parser boundary coverage
