+++ testdata/parser_result/python/module_assignment.py
@@ -1,0 +1,2 @@
+# hdr
+x = 1
