+++ testdata/parser_result/python/module_import_from.py
@@ -1,0 +1,2 @@
+# hdr
+from x import y
