+++ testdata/parser_result/python/terminal_class_and_if_suffix.py
@@ -1,0 +1,6 @@
+class GrammarTests(unittest.TestCase):
+    def test_ok(self):
+        pass
+
+if __name__ == '__main__':
+    unittest.main()
