+++ testdata/parser_result/python/terminal_if_suffix.py
@@ -1,0 +1,5 @@
+class A:
+    pass
+
+if __name__ == '__main__':
+    unittest.main()
