+++ testdata/parser_result/python/trailing_self_call.py
@@ -1,0 +1,3 @@
+    def foo():
+        x = 1;
+    foo()
